fixed radians bugs

+++ uibk_arm_controller/include/uibk_arm_controller/arm_controller.hpp
@@ -110,7 +110,7 @@ class Motor {
 		
         void setNextState(double state);
 
-        int getMaxVelLimit();
+        double getMaxVelLimit();
 	
 };
